project done

+++ app/Http/Controllers/ArticlesController.php
@@ -14,12 +14,12 @@ class ArticlesController extends Controller {
 
     public function __construct()
     {
-        $this->middleware('auth');
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
 
     public function index()
     {
-        $articles = Article::latest('published_at')->get(); //->published()
+        $articles = Article::latest('updated_at')->published()->get();
 
         return view('articles.index', compact('articles'));
     }
@@ -39,13 +39,15 @@ class ArticlesController extends Controller {
     public function store(ArticleRequest $request)
     {
         $article = Auth::user()->articles()->create($request->all());
-        $article->tags()->attach($this->tagsInput($request));
+        $this->syncTags($request, $article);
 
         return redirect()->route('articles.index')->with(['flash_info' => 'Article created successfully']);
     }
 
     public function edit(Article $article)
     {
+        $this->authorize('isRightUser', $article);
+
         $tags = $this->tagsList();
         $selected = $article->tags()->lists('id')->all();
 
@@ -54,23 +56,28 @@ class ArticlesController extends Controller {
 
     public function update(ArticleRequest $request, Article $article)
     {
+        $this->authorize('isRightUser', $article);
+
         $article->update($request->all());
-        $article->tags()->sync($this->tagsInput($request));
+        $this->syncTags($request, $article);
 
         return redirect()->route('articles.index')->with(['flash_info' => 'Article edited successfully']);
     }
 
     public function destroy(Article $article)
     {
+        $this->authorize('isRightUser', $article);
+
         $article->delete();
 
         return redirect()->route('articles.index')->with(['flash_info' => 'Article deleted']);
     }
 
-    /**
-     * Tag list to pupulate form
-     * @return mixed
-     */
+    private function syncTags($request, $article)
+    {
+        $article->tags()->sync($request->input('tags'));
+    }
+
     private function tagsList()
     {
         $tags = Tag::lists('name', 'id');
@@ -78,13 +85,4 @@ class ArticlesController extends Controller {
         return $tags;
     }
 
-    /**
-     * Tag id's from input form
-     * @param ArticleRequest $request
-     * @return array|string
-     */
-    private function tagsInput(ArticleRequest $request)
-    {
-        return $request->input('tags');
-    }
 }
